requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-anthropic
wikipedia
pydantic
openai